pyproject.toml
[tool.poetry]
name = "coursework8"
version = "0.1.0"
description = ""
authors = ["Nastya Sol <an5oldatova@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
dotenv = "^0.9.9"
psycopg2 = "^2.9.10"
celery = "^5.4.0"
pillow = "^11.1.0"
djangorestframework = "^3.15.2"
django-filter = "^25.1"
drf-yasg = "^1.21.10"
requests = "^2.32.3"
djangorestframework-simplejwt = "^5.5.0"
django-cors-headers = "^4.7.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py

)
'''

[tool.isort]
line_length = 119